Test sendDeployTx fn
